Fixing Subject to be in it's own file for better PSR-0 compatibility, and to fix an issue when being used with construct/structentities/autoloading in Drupal
